move CRssManager into seperate header/source files
