+++ xbmc/utils/RssManager.h
@@ -1,0 +1,52 @@
+#pragma once
+/*
+ *      Copyright (C) 2005-2013 Team XBMC
+ *      http://www.xbmc.org
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with XBMC; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include <vector>
+
+class CRssReader;
+class IRssObserver;
+
+class CRssManager
+{
+public:
+  CRssManager();
+  ~CRssManager();
+
+  void Start();
+  void Stop();
+  void Reset();
+  bool IsActive() const { return m_bActive; }
+
+  bool GetReader(int controlID, int windowID, IRssObserver* observer, CRssReader *&reader);
+
+private:
+  struct READERCONTROL
+  {
+    int controlID;
+    int windowID;
+    CRssReader *reader;
+  };
+
+  std::vector<READERCONTROL> m_readers;
+  bool m_bActive;
+};
+
+extern CRssManager g_rssManager;
