+++ xbmc/utils/RssReader.cpp
@@ -436,55 +436,3 @@ void CRssReader::CheckForUpdates()
 
   m_requestRefresh = false;
 }
-
-CRssManager g_rssManager;
-
-CRssManager::CRssManager()
-{
-  m_bActive = false;
-}
-
-CRssManager::~CRssManager()
-{
-  Stop();
-}
-
-void CRssManager::Start()
- {
-   m_bActive = true;
-}
-
-void CRssManager::Stop()
-{
-  m_bActive = false;
-  for (unsigned int i = 0; i < m_readers.size(); i++)
-  {
-    if (m_readers[i].reader)
-      delete m_readers[i].reader;
-  }
-  m_readers.clear();
-}
-
-// returns true if the reader doesn't need creating, false otherwise
-bool CRssManager::GetReader(int controlID, int windowID, IRssObserver* observer, CRssReader *&reader)
-{
-  // check to see if we've already created this reader
-  for (unsigned int i = 0; i < m_readers.size(); i++)
-  {
-    if (m_readers[i].controlID == controlID && m_readers[i].windowID == windowID)
-    {
-      reader = m_readers[i].reader;
-      reader->SetObserver(observer);
-      reader->UpdateObserver();
-      return true;
-    }
-  }
-  // need to create a new one
-  READERCONTROL readerControl;
-  readerControl.controlID = controlID;
-  readerControl.windowID = windowID;
-  reader = readerControl.reader = new CRssReader;
-  m_readers.push_back(readerControl);
-  return false;
-}
-
